Rename tables1 to tables in all files
